Implement workaround to recompose translation input fields on remove translation field.
