+++ app/src/main/java/com/dkat/wordbook/ui/compose/NavHost.kt
@@ -37,13 +37,6 @@ fun WordbookNavHost(
     ) {
         composable(route = Screen.Home.route) {
             HomeScreen(
-                readSource = { id: Int ->
-                    viewModel.readSource(id)
-                },
-                createWordWithTranslation = { word: Word_B, translation: Translation ->
-                    viewModel.createWordWithTranslation(word, translation)
-                },
-                sources = sources,
                 onDeleteWordItemClick = { word: Word_B ->
                     viewModel.deleteWord(word)
                 },
